Refactor prime factorization to use long long type for dividers and add implementation for summing digits of 2^1000

+++ src/3_Task/main.cpp
@@ -14,7 +14,7 @@ int main()
   // int num = 13195;
   long long num = 600851475143;
 
-  vector<int> dividers = primeFactorization(num);
+  vector<long long> dividers = primeFactorization(num);
 
   for (int d : dividers)
   {
